requirements.txt
llama-index
llama-index-llms-groq
python-dotenv
pandas
llama-index-readers-file
llama-index-embeddings-huggingface

